_thread: extract into module/_thread/ (lock + thread helpers together)
importing.rs shrinks by 393 lines.
