+++ pyre/pyre-interpreter/src/module/_thread/mod.rs
@@ -1,0 +1,8 @@
+//! _thread module — PyPy: pypy/module/thread/
+//!
+//! Single-threaded pyre: Lock / RLock state lives in the instance dict
+//! as `_locked_count`; allocate_lock / start_new_thread / etc. are
+//! stubs.
+
+pub mod interp_thread;
+pub mod moduledef;
