+++ pyre/pyre-interpreter/src/module/_thread/moduledef.rs
@@ -1,0 +1,7 @@
+//! _thread module definition — PyPy: pypy/module/thread/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_thread::register_module(ns);
+}
